+++ yampt.editor/source/editor_window.hpp
@@ -4,6 +4,8 @@
 #include <QMainWindow>
 
 class QAction;
+class QLineEdit;
+class QToolButton;
 class plugin_workspace_view_t;
 
 class editor_window_t : public QMainWindow
@@ -22,7 +24,15 @@ private:
 	void load_config();
 	void save_config();
 	void on_open_settings();
+	void on_search_apply();
+	void on_search_clear();
 
 	plugin_workspace_view_t * m_plugin_workspace_view = nullptr;
 	settings_store_t m_settings { "yEditor.ini" };
+
+	QLineEdit * m_search_field = nullptr;
+	QToolButton * m_case_sensitive_btn = nullptr;
+	QToolButton * m_regex_btn = nullptr;
+	QToolButton * m_search_id_btn = nullptr;
+	QToolButton * m_search_name_btn = nullptr;
 };